api automation tests around profile endpoint
